refactor: introduce @bitwarden/serialization
