+++ .github/CODEOWNERS
@@ -96,6 +96,7 @@ libs/logging @bitwarden/team-platform-dev
 libs/storage-test-utils @bitwarden/team-platform-dev
 libs/messaging @bitwarden/team-platform-dev
 libs/messaging-internal @bitwarden/team-platform-dev
+libs/serialization @bitwarden/team-platform-dev
 # Web utils used across app and connectors
 apps/web/src/utils/ @bitwarden/team-platform-dev
 # Web core and shared files
