+++ libs/serialization/README.md
@@ -1,0 +1,5 @@
+# serialization
+
+Owned by: platform
+
+Core serialization utilities
