+++ libs/serialization/project.json
@@ -1,0 +1,33 @@
+{
+  "name": "serialization",
+  "$schema": "../../node_modules/nx/schemas/project-schema.json",
+  "sourceRoot": "libs/serialization/src",
+  "projectType": "library",
+  "tags": [],
+  "targets": {
+    "build": {
+      "executor": "@nx/js:tsc",
+      "outputs": ["{options.outputPath}"],
+      "options": {
+        "outputPath": "dist/libs/serialization",
+        "main": "libs/serialization/src/index.ts",
+        "tsConfig": "libs/serialization/tsconfig.lib.json",
+        "assets": ["libs/serialization/*.md"]
+      }
+    },
+    "lint": {
+      "executor": "@nx/eslint:lint",
+      "outputs": ["{options.outputFile}"],
+      "options": {
+        "lintFilePatterns": ["libs/serialization/**/*.ts"]
+      }
+    },
+    "test": {
+      "executor": "@nx/jest:jest",
+      "outputs": ["{workspaceRoot}/coverage/{projectRoot}"],
+      "options": {
+        "jestConfig": "libs/serialization/jest.config.js"
+      }
+    }
+  }
+}
