+++ libs/serialization/src/serialization.spec.ts
@@ -1,0 +1,8 @@
+import * as lib from "./index";
+
+describe("serialization", () => {
+  // This test will fail until something is exported from index.ts
+  it("should work", () => {
+    expect(lib).toBeDefined();
+  });
+});
